api/pyproject.toml
[project]
name = "mandev-api"
version = "0.1.0"
description = "FastAPI backend for man.dev"
requires-python = ">=3.12"
dependencies = [
    "mandev-core",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "piccolo[postgres]>=1.0",
    "asyncpg>=0.30",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "pydantic-settings>=2.0",
    "email-validator>=2.0",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "anyio>=4.0",
    "httpx>=0.28",
    "piccolo[sqlite]>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mandev_api"]
